change data type

+++ entities/farm.go
@@ -15,7 +15,7 @@ type Farm struct {
 	Description             string
 	Picture                 string
 	MinimumInvestmentAmount float64
-	CountInvestment         uint64
+	CountInvestment         float64
 	CreatedAt               time.Time
 	UpdatedAt               time.Time
 	DeletedAt               gorm.DeletedAt
